Number and string Collection

+++ 0x04-more_functions_nested_loops/9-fizz_buzz.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	int count;
+	short count;
 
 	for (count = 1; count <= 100; count++)
 	{
